+++ gameserver/realm/commands/command/AdminCommands.cs
@@ -35,8 +35,10 @@ namespace LoESoft.GameServer.realm.commands
             {
                 string name = string.Join(" ", args.Skip(1).ToArray());
                 //creates a new case insensitive dictionary based on the XmlDatas
-                Dictionary<string, int> icdatas = new Dictionary<string, int>(GameServer.Manager.GameData.IdToObjectType, StringComparer.OrdinalIgnoreCase);
-                if (!icdatas.TryGetValue(name, out int objType) ||
+                Dictionary<string, ushort> icdatas = new Dictionary<string, ushort>(
+                    GameServer.Manager.GameData.IdToObjectType,
+                    StringComparer.OrdinalIgnoreCase);
+                if (!icdatas.TryGetValue(name, out ushort objType) ||
                     !GameServer.Manager.GameData.ObjectDescs.ContainsKey(objType))
                 {
                     player.SendInfo("Unknown entity!");
@@ -55,8 +57,10 @@ namespace LoESoft.GameServer.realm.commands
             {
                 string name = string.Join(" ", args);
                 //creates a new case insensitive dictionary based on the XmlDatas
-                Dictionary<string, int> icdatas = new Dictionary<string, int>(GameServer.Manager.GameData.IdToObjectType, StringComparer.OrdinalIgnoreCase);
-                if (!icdatas.TryGetValue(name, out int objType) ||
+                Dictionary<string, ushort> icdatas = new Dictionary<string, ushort>(
+                    GameServer.Manager.GameData.IdToObjectType,
+                    StringComparer.OrdinalIgnoreCase);
+                if (!icdatas.TryGetValue(name, out ushort objType) ||
                     !GameServer.Manager.GameData.ObjectDescs.ContainsKey(objType))
                 {
                     player.SendHelp("Usage: /spawn <entityname>");
@@ -158,9 +162,9 @@ namespace LoESoft.GameServer.realm.commands
 
             try
             {
-                Dictionary<string, int> icdatas = new Dictionary<string, int>(GameServer.Manager.GameData.IdToObjectType, StringComparer.OrdinalIgnoreCase);
+                Dictionary<string, ushort> icdatas = new Dictionary<string, ushort>(GameServer.Manager.GameData.IdToObjectType, StringComparer.OrdinalIgnoreCase);
 
-                if (!icdatas.TryGetValue(name, out int objType))
+                if (!icdatas.TryGetValue(name, out ushort objType))
                 {
                     player.SendError("Unknown type!");
                     return false;
